Refactor runtime assembly and tail collaborators
