+++ app/service/navigator_runtime/tail_components/telemetry.py
@@ -7,6 +7,8 @@ from navigator.core.service.scope import profile
 from navigator.core.telemetry import LogCode, Telemetry, TelemetryChannel
 from navigator.core.value.message import Scope
 
+from .edit_request import TailEditDescription
+
 
 class TailTelemetry:
     """Emit structured telemetry for navigator tail operations."""
@@ -27,6 +29,28 @@ class TailTelemetry:
             method=method,
             scope=self._profile,
             **fields,
+        )
+
+    def record_get(self) -> None:
+        """Record the retrieval of the last navigator message."""
+
+        self.emit("last.get")
+
+    def record_delete(self) -> None:
+        """Record deletion of the last navigator message."""
+
+        self.emit("last.delete")
+
+    def record_edit(self, description: TailEditDescription) -> None:
+        """Record metadata about an edit applied to the navigator tail."""
+
+        self.emit(
+            "last.edit",
+            payload={
+                "text": description.text,
+                "media": description.media,
+                "group": description.group,
+            },
         )
